feat: exported user settings components for avatar upload, email change, password change, username update, and account deletion from settings-from
